[PackageBuilder] add NeonLoaderAwareKernelTrait
