+++ packages/PackageBuilder/tests/Adapter/Symfony/AppKernel.php
@@ -7,9 +7,12 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\BundleInterface;
 use Symfony\Component\HttpKernel\Kernel;
 use Symplify\PackageBuilder\Adapter\Symfony\Parameter\ParameterProvider;
+use Symplify\PackageBuilder\Neon\NeonLoaderAwareKernelTrait;
 
 final class AppKernel extends Kernel
 {
+    use NeonLoaderAwareKernelTrait;
+
     /**
      * @var string
      */
