+++ packages/PackageBuilder/tests/Adapter/Symfony/Parameter/ParameterProviderSource/Neon/included-config.neon
@@ -1,0 +1,2 @@
+parameters:
+    one: 1
