+++ packages/PackageBuilder/tests/Adapter/Symfony/Parameter/ParameterProviderSource/Neon/including-config.neon
@@ -1,0 +1,5 @@
+includes:
+    - included-config.neon
+
+parameters:
+    two: 2
